moved classes to python_timer_stuff/__init__.py
